requirements.txt
altair==4.1.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
beautifulsoup4==4.8.2
bleach==3.1.4
blinker==1.4
boto3==1.12.35
botocore==1.15.35
bs4==0.0.1
cdqa==1.3.9
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fake-useragent==0.1.11
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.13.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.2
pandas==0.25.0
pandocfilters==1.4.2
parso==0.6.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pycparser==2.20
pydeck==0.3.0b4
Pygments==2.6.1
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.3
pyzmq==19.0.0
regex==2020.2.20
requests==2.22.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.21.2
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.85
six==1.14.0
soupsieve==2.0
streamlit==0.56.0
terminado==0.8.3
testpath==0.4.4
tika==1.19
toml==0.10.0
toolz==0.10.0
torch==1.2.0
tornado==5.1.1
tqdm==4.32.2
traitlets==4.3.3
transformers==2.1.1
tzlocal==2.0.0
urllib3==1.25.8
validators==0.14.3
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
zipp==3.1.0
